LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/merge-timeline-offset-keyframes-expected.txt


FAIL Keyframes with same easing and timeline offset are merged. assert_equals: number of frames:  expected 4 but got 0
FAIL Keyframes with same timeline offset but different easing function are not merged. promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'anim.ready')"
